refactor: reorganize and define MessageData interface in types.ts
